Rollback 041ec4d
The spacer <div> is required to minimize the <main> height to just fit its content. Without the spacer, <main> will consume the entire screen height and spread its children across the entire height. When there are not enough cards to fill the screen height, this spread results in ugly spacing.