Remove canonical URL header
It causes issues with SEO when the search engine's scraper doesn't finish running the JS to update the canonical URL for an article. There are also no alternate URLs so the canonical URL is not useful anyway (the canonical URL was originally intended to de-duplicate the "www." subdomain but "www." already redirects to the root).